requirements.txt
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.20.0
six==1.16.0
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.8
python-dotenv==0.20.0
